gtpu: encap vrf id fixed
The encapsulated (outer) VRF id can be explicitly set to a different
value than the interface VRF id. This allows the tunneled traffic
to be routed by a separate FIB table.
This is necessary because the VPP gtpu plugin implements a check
to verify whether the ingress interface vrf assignment matches
the encap vrf id of the tunnel. If not, the packet is dropped.
Without a different VRF value, it would not be possible to implement
an overlay network with separate routing tables.

Dependencies for the GTPU VRF id and multicast interface are added.

The VRF id is inherited from the unnumbered interface only if
the value is not explicitly set to non-0. This allows the tunnel
interface to re-use the IP address of another interface from a different
VRF table.