feat: GTP-U tunnel support (#1560)
* interface plugin: GTP-U tunnel support added

Initial implementation of GTP-U tunnel interface configuration added.
Currently VPP 20.01-324 and 20.01-379 are supported only.

* gtpu: Code review comments addressed

- GTPU tunnel src/dst addresses are validated before writing into
the kvs

- if-else in dump interface call is restructured.

* gtpu: encap vrf id fixed

The encapsulated (outer) VRF id can be explicitly set to a different
value than the interface VRF id. This allows the tunneled traffic
to be routed by a separate FIB table.
This is necessary because the VPP gtpu plugin implements a check
to verify whether the ingress interface vrf assignment matches
the encap vrf id of the tunnel. If not, the packet is dropped.
Without a different VRF value, it would not be possible to implement
an overlay network with separate routing tables.

Dependencies for the GTPU VRF id and multicast interface are added.

The VRF id is inherited from the unnumbered interface only if
the value is not explicitly set to non-0. This allows the tunnel
interface to re-use the IP address of another interface from a different
VRF table.

* gtpu: source code formatted

* licence note added

* gtpu: vppcalls test cases added

IP address conversion fixed: an IPv4 address was successfully
parsed as IPv6 by mistake.

* gtpu: added support for vpp 19.08

* gtpu: added support for vpp 19.04

* gtpu: added test cases for dump interface function

* gtpu: added integration tests

These tests depend on the VPP bugfix bd0a00a45637c1dde533ef6c3798418c2ab15009

* gtpu: revert the change with the VRF overwrite

This change could affect other interfaces as well, better to revert,
even if VPP supports this case.

The scenario can be made work with configuring a loopback interface
with the desired VRF id and using that interface as the tunnel
unnumbered int.

* gtpu: disable interface dump in the broken VPP versions

Disable the dump gtpu interface API call in the gtpu integration
test with some of the VPP versions that do not include the fix yet.

* gtpu: golint fix, renamed dumpApiOk to dumpAPIOk

* gtpu: fixed gtpu implementation after rebase on dev

- updated imports to modules
- fixed remaining conflicts
- regenerated proto files