Fixed unnumbered references to discovered interfaces
The "UNTAGGED-" prefix and setting the name of the discovered interfaces
without a tag is moved to the interface implementation.

It is too late to set the name in the desciptor because the dump
interfaces function already rely on that:
When any other interface is unnumbered and references this target interface
without a name, the vrf and ip address values cannot be later resolved.