feat: Allow AF-PACKET to reference target Linux interface via logical name
Added <linux_interface> field to AF-PACKET link definition, which
allows to reference the target Linux interface via logical name
as opposed to its host name. Previously the physical reference
was chosen over the logical because it was not possible to reference
Linux interfaces not managed by the same vpp-agent, but with the
introduction of the EXISTING Linux interface type this is no longer
the case.

Signed-off-by: Milan Lenco <milan.lenco@pantheon.tech>