Define DefaultNat44GlobalConfig for NAT handler
This allows having different default configuration for each VPP version.
In VPP 20.01 the virtual reassembly has been moved from NAT to IP plugin and
can be configured for both IPv4/IPv6. However due to incompatibilities, like
Timeout being changed from sec to ms, it is disabled for now.

The IP Reassembly configuration should be implemented in L3 plugin as
it is not NAT-specific feature in any way.

Signed-off-by: Ondrej Fabry <ofabry@cisco.com>