test: Run VPP+VPPAgent in a container for e2e tests
This is a change that affects how VPP and VPPAgent are started for end-to-end tests.
Until now both VPP and VPPAgent processes were started as a processes in the system,
which was a special built container for running end-to-end tests. However to allow
for multi-VPP tests and make the execution of tests closer match the real use it is
better to run VPP+VPPAgent using a separate container that is normally built.

Signed-off-by: Ondrej Fabry <ofabry@cisco.com>